Model event store flavors using feature variants
