+++ account-service/src/main/java/lt/rieske/accounts/eventstore/Configuration.java
@@ -8,8 +8,9 @@ import lt.rieske.accounts.eventsourcing.EventStore;
 import lt.rieske.accounts.infrastructure.TracingConfiguration;
 
 import javax.sql.DataSource;
+import java.lang.reflect.InvocationTargetException;
+import java.lang.reflect.Method;
 import java.util.function.Function;
-
 
 public class Configuration {
 
@@ -22,12 +23,20 @@ public class Configuration {
         return blobEventStore(jdbcUrl, username, password, ds -> pooledMeteredDataSource(tracingConfiguration.decorate(ds), meterRegistry));
     }
 
-    static BlobEventStore blobEventStore(String jdbcUrl, String username, String password,
+    private static BlobEventStore blobEventStore(String jdbcUrl, String username, String password,
                                          Function<DataSource, DataSource> initializer) {
-        if (jdbcUrl.startsWith("jdbc:postgresql://")) {
-            return PostgresEventStoreFactory.postgresEventStore(jdbcUrl, username, password, initializer);
-        } else if (jdbcUrl.startsWith("jdbc:mysql://")) {
-            return MySqlEventStoreFactory.mysqlEventStore(jdbcUrl, username, password, initializer);
+        try {
+            if (jdbcUrl.startsWith("jdbc:postgresql://")) {
+                Class<?> eventstoreFactoryClass = Class.forName("lt.rieske.accounts.eventstore.EventStoreFactory");
+                Method initMethod = eventstoreFactoryClass.getMethod("makeEventStore", String.class, String.class, String.class, Function.class);
+                return (BlobEventStore) initMethod.invoke(eventstoreFactoryClass, jdbcUrl, username, password, initializer);
+            } else if (jdbcUrl.startsWith("jdbc:mysql://")) {
+                Class<?> eventstoreFactoryClass = Class.forName("lt.rieske.accounts.eventstore.EventStoreFactory");
+                Method initMethod = eventstoreFactoryClass.getMethod("makeEventStore", String.class, String.class, String.class, Function.class);
+                return (BlobEventStore) initMethod.invoke(eventstoreFactoryClass, jdbcUrl, username, password, initializer);
+            }
+        } catch (ClassNotFoundException | NoSuchMethodException | IllegalAccessException | InvocationTargetException e) {
+            throw new IllegalStateException(e);
         }
         throw new IllegalStateException(String.format("Unsupported JDBC URL '%s'", jdbcUrl));
     }
