+++ account-service/src/test/java/lt/rieske/accounts/eventstore/H2.java
@@ -3,6 +3,7 @@ package lt.rieske.accounts.eventstore;
 import org.h2.jdbcx.JdbcDataSource;
 
 import javax.sql.DataSource;
+import java.util.function.Function;
 
 public class H2 {
 
@@ -14,6 +15,10 @@ public class H2 {
 
     public static H2 postgres() {
         return new H2("PostgreSQL");
+    }
+
+    public static BlobEventStore postgresEventStore() {
+        return EventStoreFactory.postgresEventStore(postgres().dataSource(), Function.identity());
     }
 
     private H2(String mode) {
