+++ eventstore/src/mysql/resources/db/migration/V001__init.sql
