+++ eventstore/src/postgres/resources/db/migration/V001__init.sql
