--- gradle/build-logic/src/main/groovy/account.integration-tests.gradle
@@ -1,20 +1,0 @@
-sourceSets {
-    integrationTest {
-        compileClasspath += sourceSets.test.output
-        runtimeClasspath += sourceSets.test.output
-    }
-}
-
-configurations {
-    integrationTestImplementation.extendsFrom testImplementation
-    integrationTestRuntimeOnly.extendsFrom testRuntimeOnly
-}
-
-def integrationTest = tasks.register('integrationTest', Test) {
-    it.shouldRunAfter(test)
-
-    it.testClassesDirs = sourceSets.integrationTest.output.classesDirs
-    it.classpath = sourceSets.integrationTest.runtimeClasspath
-}
-
-tasks.named('check') { it.dependsOn(integrationTest) }
