+++ gradle/build-logic/src/main/groovy/account.java-conventions.gradle
@@ -1,8 +1,7 @@
 plugins {
-    id "java"
-    id "java-test-fixtures"
-    id "jacoco"
-    id "account.integration-tests"
+    id("java")
+    id("java-test-fixtures")
+    id("jacoco")
 }
 
 java {
@@ -21,7 +20,7 @@ tasks.withType(JavaCompile).configureEach {
 
 tasks.withType(Test).configureEach {
     useJUnitPlatform()
-    maxParallelForks = 4
+    maxParallelForks = 8
     jvmArgs += "--enable-preview"
 }
 
@@ -43,7 +42,7 @@ jacoco {
 
 jacocoTestReport {
     reports {
-        executionData.setFrom project.fileTree(dir: '.', include: '**/build/jacoco/*.exec')
+        executionData.setFrom project.fileTree(dir: ".", include: "**/build/jacoco/*.exec")
         xml.enabled true
     }
 }
