+++ gradle/build-logic/src/main/groovy/account.service-conventions.gradle
@@ -1,19 +1,19 @@
 plugins {
-    id "account.java-conventions"
-    id "application"
+    id("account.java-conventions")
+    id("application")
 }
 
 sourceSets {
     e2eTest
 }
 
-def e2eTest = tasks.register('e2eTest', Test) {
-    it.dependsOn("assemble", "integrationTest")
+def e2eTest = tasks.register("e2eTest", Test) {
+    it.dependsOn("assemble")
 
     it.testClassesDirs = sourceSets.e2eTest.output.classesDirs
     it.classpath = sourceSets.e2eTest.runtimeClasspath
 }
 
-tasks.named('check') { it.dependsOn("test", e2eTest) }
+tasks.named("check") { it.dependsOn("test", e2eTest) }
