+++ settings.gradle
@@ -2,5 +2,3 @@ rootProject.name = "account"
 includeBuild("gradle/build-logic")
 include("account-service")
 include("eventstore")
-include("eventstore-postgres")
-include("eventstore-mysql")
